+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure continuous tasks stealing on idle start at worker node idling
+
 ## [3.0.7] - 2023-11-24
 
 ### Changed